Varying max presentations per trial, changing parameter values, and fixing WM match code.
Previously we had no way of specifying on a per simulation basis what
the max number of presentations for a given trial - it was fixed at 20 -
but this put a limit on the minimum # of unique items we could use (#
unique items >= max presentations per trial) so we modified this to be
able to create ultra high PI conditions.  Also, modified general
parameters in search of results that would give behavior expected of
model.  Lastly, WM had previously been useless after the initial
encoding because it's match value calculation was using the WM encoding
noise parameter which is 0, but the post-rehearsal, retrieved-from-EM
memory was generated with a non-zero noise parameter.  Modified it so
that the match calculations accept the noise parameter and used this
throughout REMplusWM.