Fix WM Rehearsal Match Strength Tiebreakers Case
Previous version of code did not take into account the possibility that
WM rehearsal process might encounter multiple EM traces with the same
exact, highest match value.  In this scenario, the max function simply
returned the first index of the maximum value, which necessarily
excludes the last EM trace encoded, thus blocking a successful retrieval
in any of the tiebreaker cases.

Modified logic in REMplusWM.updateWMifNeeded to look for tiebreaker case
and automatically select the correct EM trace if it is one of the
highest match values.  If the correct EM trace is not among the maximum
values, then we randomly select one (note: previous behavior always
chose the first value).  Also added a function that works similar to the
max function in matlab, but returns multiple indices if there are
multiple max values.

The modification to the tiebreaker case affects WM is specific, though
potentially impactful.  We might want to change the behavior to always
randomly select amongst tiebreakers (instead of selecting the correct EM
trace if it's among the max values)