Major Bug Fix WM Match Calculation
Bug: When matching EM traces to WM trace for purpose of rehearsal, would
get overabundance of failed rehearsals with even the slightest amount of
WM trace decay.  This happened because the initial version of
calculateItemTracesOddsRatio was meant for comparing items and memory
traces, assuming that items never had features which were 0.  However,
co-opting this function to compare WM and EM traces to each other failed
to look for 0 entries in the WM trace, thus any 0-valued feature in the
WM trace was causing an unconditional mismatch between EM and WM even
though they might belong to the same item.

Modify calculateItemTracesOddsRatio to account for 0s in either of the
two entities being compared by only comparing entries that are non-zero
in both.