Convenience variables for aggregating sim results
Sometimes we want to give a description of our aggregated results, but
sometimes we don't want the script to wait for our description (if we
have to leave the computer or something).  Additionally sometimes we
want to automatically plot the AUC curves after we've aggregated this.

Add variables PLOT_AUC and NO_DESCRIPTION which control whether or not
we call plotAUCEMvsWMAggResults and wait for a description respectively.