Major Bug Fix AUC plotting
Previously the plotAUC family of functions assumed that each trial's
WMpastStrengths (or EM) variables, in a Trial_Simulator object,
consisted only of the values for that trial.  Under this assumption,
generating the AUC curve would append the past strengths for each trial.
In actuality, the pastStrength variables for a trial hold ALL the
pastStrength values from the current trial backwards.  Thus the plot AUC
functions, by appending these variables for each trial, were repeating
multiple trials each trial so that trial 100 would have 100 copies of
Trial 1, 99 copies of Trial 2, etc.

Modify the plotAUC functions so that we use the
EM/WMpresentationStrengthsPerTrial variables which are only contain the
current trial information - thus yielding sensible results when
appending.