Running in current matlab session on cluster
Sometimes we want to run things in the current matlab session even when
we're on the cluster - previously we only ran things in the current
matlab session if we were not on the cluster.

Add a function to PM_task.m runInCurrentMatlabEvenIfOnCluster that will
launch a script and NOT exit after it's done.  This is accomplished by
setting a shouldExitIfOnCluster variable which is checked by the family
of runSim*.m scripts.

All runSim scripts will exitIfOnCluster unless the variable
shouldExitIfOnCluster is set to false.