Add threshold-based rejection of WM rehearsals
Some models of memory (such as SAM) have two stages of recall: sampling
that locates the memory and then recovery which retrieves the details of
the memory.  Thus, we'd like to explore the possibility that WM
rehearsal might return a pretty crappy match (kind of like the sampling
phase), but fail to get stored in WM (kinda like recovery) because the
match strength is so low.  We think this might help us out with our
difficulties in getting the WM system to interact with the PI
manipulation.

We create an additional option,minimumRetrievedLogStrengthWM in
REMplusWM.m (which can be set in Trial_Simulator.m) that will reject a
retrieved trace if it is below this threshhold, leaving the WM trace in
whatever state it was before the rehearsal attempt.
	Add machinery to keep track of when our model rejected a WM
rehearsal in Trial_Simulator.m
	Add machinery to keep track of best matching memory strength for
each WM rehearsal attempt and plotting scripts to look at the average
memory strength for a rehearsal given that it returned either the right
Item with the wrong context, a wrong item altogether or the correct
item/context combo.

The only modified functionality is that
plotRehearsalFailuresOverTrials.m will now plot rejected trials in
green, so that it technically does not count rejected rehearsals as
failed rehearsals.